requirements.txt
Flask
flask-socketio
eventlet
python-chess